split into legacy script
